backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth-oauthlib
google-api-python-client
langchain
langchain-google-genai
langgraph
itsdangerous
pydantic
requests
starlette
httpx
psycopg2-binary